requirements 2.txt
streamlit
alpaca-trade-api
scikit-learn
pandas
ta
pyyaml